refactor(JwtAuthenticationFilter) : JWT 인증 필터 로직 개선
JWT 검증 및 사용자 인증 로직을 분리하여 JwtTokenVerifier를 활용하도록 변경.
Optional을 사용하여 토큰 추출 및 검증을 간결하게 처리하고, 인증 객체 생성을 UserDetailsFactory로 위임하여 책임을 명확히 분리함.
